Extract mobile performance modal, add to performance confirmations hook
